write a global test for wrong center numbers
